+++ src/main/java/com/ll/gooHaeYu/domain/member/member/service/MemberService.java
@@ -6,7 +6,7 @@ import com.ll.gooHaeYu.domain.member.member.entity.Member;
 import com.ll.gooHaeYu.domain.member.member.repository.MemberRepository;
 import com.ll.gooHaeYu.global.exception.CustomException;
 import com.ll.gooHaeYu.global.exception.ErrorCode;
-import com.ll.gooHaeYu.standard.base.utils.JwtUtil;
+import com.ll.gooHaeYu.standard.base.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
